order_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
kafka-python>=2.0.2
python-jose[cryptography]>=3.3.0
